Move progress lib to output folder 4 consistency
