+++ lib/main-loop.js
@@ -8,7 +8,7 @@
 
 require('dotenv').config(); // Load environment variables from .env file
 const async = require('async');
-const progress = require('./progress');
+const progress = require('./output/progress');
 const logging = require('./output/logging');
 const csvReader = require('./input/csv-file-reader');
 const cloudinary = require('cloudinary').v2;
