+++ lib/output/progress.js
